mcp_servers/rfp/requirements.txt
openai>=1.68.2
pypdf2>=3.0.1
python-dotenv>=1.0.1
tqdm